Fix compiler-specific weirdnesses in my fixes for rounding errors; patch
* applet.c: (gnome_volume_applet_scroll),
(gnome_volume_applet_key), (gnome_volume_applet_adjust_volume),
(cb_volume):
* applet.h:
Fix compiler-specific weirdnesses in my fixes for rounding errors;
patch contributed by Kazuki Shimura <kazuki@ma.ccnw.ne.jp>. Fixes
bug #316922.

+++ mixer/applet.h
@@ -92,7 +92,7 @@ typedef struct _GnomeVolumeAppletClass {
 
 void     gnome_volume_applet_adjust_volume (GstMixer      *mixer,
 					    GstMixerTrack *track,
-					    float          volume);
+					    gdouble        volume);
 GType    gnome_volume_applet_get_type (void);
 gboolean gnome_volume_applet_setup    (GnomeVolumeApplet *applet,
 				       GList             *elements);
